Replaced `named-clause' by `name-clause' to conform to the HyperSpec.
Modified the parsing code so that all clauses are parsed with no
concern about the order.  Later, we need to check the order and
generate appropriate error messages if they are out of order.
Modified the definition of the class loop-body to contain this new
list of all the clauses.  Introduced a function parse-any-clause that
parses any clause.

Modified the code generation to take the new list of all clauses into
account.

Removed the function parse-variable-clause, because it is no longer
needed with the new way of parsing clauses.

Removed the function parse-main-clause, because it is no longer needed
with the new way of parsing clauses.

Added a comment about the HyperSpec definition of the conditional
clause.  Renamed existing function `parse-conditional' to
`parse-conditional-remaining' and introduced a "normal" parser called
`parse-conditional' that just does an alternative parse of if/when and
unless.

Added a comment about the HyperSpec definition of the termination-test
clause.  Renamed existing function `parse-termination-test' to
`parse-termination-test-remaining' and introduced a "normal" parser
called `parse-termination-test' that just does an alternative parse of
if/when and unless.

All tests pass (there are not very many, but still).