Fix unbound *it-var* error
The it clause parsers will capture a clause with "it" in the form
position even when loop does not currently have it bound. To
allow for possibility that the "it" is actually a variable
reference I've made the it clause classes subclasses of the
corresponding form clauses and passed control back to the
appropriate body-form method when *it-var* is not bound.

Also, I made the singleton matcher for keyword return the token
not the matched keyword so that the form clauses could access
the original package of the "it" symbol.